Fix errors recently introduced. (#1448)
* Fix bad references on tests for BigInt TypedArrays

* Remove bad conversions for BigInt TypedArray

* Cleanup the BigInt TypedArray harness file

Remove non used code (testBigIntTypedArrayConversions)

Move the constructors list to inside the exposed function, this prevents early implementations to fail before the function is called.

* Fix bad references in TypedArrays.of (BigInt)

* Remove BigInt tests from typedarray harness test

* Use BigInt for BigInt typedArrays

* Apply last fixings on BigInt TypedArray tests

* Apply fixes to last revision from @anba

+++ test/built-ins/TypedArray/prototype/slice/BigInt/detached-buffer-zero-count-custom-ctor-other-targettype.js
@@ -27,7 +27,7 @@ testWithBigIntTypedArrayConstructors(function(TA) {
   var sample, result, other;
   var ctor = {};
   ctor[Symbol.species] = function(count) {
-    other = TA === Int8Array ? Int16Array : Int8Array;
+    other = TA === BigInt64Array ? BigUint64Array : BigInt64Array;
     $DETACHBUFFER(sample.buffer);
     return new other(count);
   };
